Added multi point collision detection.

+++ pbrenderer/LYCudaHelper.cuh
@@ -22,8 +22,8 @@ namespace LYCudaHelper
 	void freeArray(void *devPtr);
 	void *mapGLBufferObject(struct cudaGraphicsResource **cuda_vbo_resource);
 	void unmapGLBufferObject(struct cudaGraphicsResource *cuda_vbo_resource);
-	void copyArrayFromDevice(void *host, const void *device, struct cudaGraphicsResource **cuda_vbo_resource, int size);
-	void copyArrayToDevice(void *device, const void *host, int offset, int size);
+	void copyArrayFromDevice(void *host, const void *device, struct cudaGraphicsResource **cuda_vbo_resource, size_t size);
+	void copyArrayToDevice(void *device, const void *host, int offset, size_t size);
 	void registerGLBufferObject(uint vbo, struct cudaGraphicsResource **cuda_vbo_resource);
 	void unregisterGLBufferObject(struct cudaGraphicsResource *cuda_vbo_resource);
 	void getMappedPointer(void **device, void *host, uint flag = 0);
